add disclaimer class; make LastUpdated dynamic
LastUpdated is the notification on the top of the page. I used to update it manually because I had created a simple SPA, but since that broke the browser's history I switched to just HTML pages. So now it's dynamic and it'll update if the page updates. Not a perfect solution, since I kinda want it updated when I update the <main> content, not just anything. Oh well.