add Cache Busting option
By adding "?v=1"  to the css links, I can force an update of the cache on the client side by updating the number. This is apparently called Cache Busting. Now I could set the expiry date of the CSS to 10 years into the future if I wanted. Not that I can, because I'm currently hosted on github.io :(